refactor(data): unify DataHandle→record mapping in data_record_mapper
The success path of DefaultStepExecutor.executeModelMethod had ~70
inline lines mapping DataHandle to DataRecord (for resources) or
FileDataRecord (for files). PR #1252 review flagged that this site
was "out of step" with the canonical mappers c91a22fb consolidated:
fromRow (CatalogRow→DataRecord) and fromData (Data→DataRecord). On
inspection there was no unified handle→record mapper — so this commit
adds two new functions in the canonical location:

- fromResourceHandle(handle, modelType, modelId, fallbackModelName, dataRepo)
  → DataRecord. Reads JSON content for application/json into
  attributes; otherwise empty. content="" by design (downstream CEL
  references attributes.<field>, not raw text).

- fromFileHandle(handle, modelType, modelId, dataRepo)
  → FileDataRecord | null. Stats the on-disk content path for size.
  Returns null when the file is missing (rare race with cleanup).

handleMethodSuccess now calls these instead of inlining the work. The
DataRecord shape is owned by one module: future field additions
update the mappers, not N inline sites.

FileDataRecord moved from src/domain/expressions/model_resolver.ts
to src/domain/data/data_record.ts (its proper home alongside
DataRecord). Old location re-exports it for back-compat. Without
this move, importing FileDataRecord into data_record_mapper.ts
would have created a new mutual dependency between data and
expressions bounded contexts (caught by integration/architecture_boundary_test.ts).

Behaviour-preserving:
- Same isLatest=true (handle was just produced)
- Same JSON-only attribute parsing for resources
- Same Deno.stat for files; null on missing
- Same dataType="resource" default vs fromData's ""
- Same modelName fallback to evaluatedDefinition.name
- Verified: deno check / lint / fmt / test (5045 passed),
  /tmp/swamp-verification check OK against baseline,
  architecture_boundary ratchet stable at 15
